+++ CHANGELOG.md
@@ -16,6 +16,7 @@ reasonable).
 - [Add Parsely tags as a fallback metadata source](https://github.com/mozilla/readability/pull/865)
 - [Fix the case that jsonld parse process is ignored when context url include the trailing slash](https://github.com/mozilla/readability/pull/833)
 - [Fixed situations where short paragraphs of legitimate content would be excluded](https://github.com/mozilla/readability/pull/867)
+- [Add `citation`, `prism`, and more `dc` metadata](https://github.com/mozilla/readability/pull/871)
 
 ## [0.5.0] - 2023-12-15